+++ src/Autocomplete/Autocomplete.stories.tsx
@@ -384,7 +384,7 @@ export const ExampleAutocompleteOpen = {
   },
   play: async ({ canvasElement }: { canvasElement: HTMLElement }) => {
     const canvas = within(canvasElement);
-    await userEvent.click(canvas.getByRole('button'));
+    await userEvent.click(canvas.getByRole('button', { name: 'Open' }));
   },
 };