+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added the `TreeView` component.
+- Adding chromatic PR worklfow snapshot tests
 
 ### Fixed
 - Fixed an issue in Static DatePicker (MUI) where the "Today" action did not correctly update or reflect the selected date when triggered from the year/month view.